Correct build for cross-compilation
The build target of the build script is not the same as the build target of the
final lib. Using the *runtime* environment variable (as opposed to the cfg var
or `env!()`) determines the actual target of the library.

For example, if you are running `cargo build --target=x86_64-apple-darwin` on
Linux then the target for the `build.rs` script *must* be linux -- because the
script is executed on the current machine -- but the target of the `sysinfo`
lib is the `--target` parameter.