Add apple-sandbox feature as a subset of apple-app-store
A common misnomer on macOS is that using the macOS sandbox means that
the App Store is also involved, which isn't always true.

For the sake of simplicity, don't yet switch macOS component reading
features to this flag to avoid a mess of feature flag requirements
internally. It is *most likely* that an application needing this data
woulnd't run inside of a sandbox anyway.