linux: Drop fishy shortcut for getting uid/gid
Getting the UID/GID by looking at the ownership of the relevant /proc
file of a process is strange. The UID/GID are provided in the status
file itself that we read anyway if this method fails. Let's always take
the UID/GID from the process' status file.

This also paves way for reliably getting the effective UID/GID of a
process, that we will be adding in a following commit.