Add support for total accumulated process CPU usage
Most, if not all, CPU usage accounting for processes provides values that count
from the creation of the process. This total value is useful for a variety of
accounting tasks beyond the snapshot value that is currently available in
sysinfo. This change adds a `fn total_cpu_usage` to `trait ProcessExt` to
provide that value.