Use `OsString` everywhere
This makes it so `OsString` / `OsStr` are used everywhere instead of
`String` / `str`, because the operating systems make no guarantee that
the values provided are valid UTF-8. The `OsString` / `OsStr` types are
specifically provided by the standard library for these situations. This
fixes situations where `sysinfo` either had to lossily lose some of the
information, had some subtle undefined behavior, and partially rejected
processes and other resources entirely, making them impossible to be
accessed by the user of this crate.