Use `OsString` for process names
Previously the stat file on Linux was parsed under the assumption that
it contains valid UTF-8. This is not necessarily the case for two
reasons:
1. Process names don't need to be valid UTF-8 at all on Linux. They can
   contain invalid bytes.
2. Even if the process name was valid UTF-8, the assumption breaks
   because the process name is limited to 15 characters, which can cut a
   valid code point in half.

This causes the stat file to not be parseable, causing the entire
process not to show up in the list. So even if you aren't using the name
of the process at all, you can't find it as part of the list.

One solution is to lossily convert the process name to a string. However
this means that the [Unicode replacement character
`�`](https://www.fileformat.info/info/unicode/char/fffd/index.htm) is
now part of those process names. The character when encoded as UTF-8 is
3 bytes. This means that process names can now be longer than 15 bytes,
or in other words, doing a name based comparison on the first 15 bytes,
such as suggested in the documentation of the crate, is no longer easily
possible.

The solution is to just provide the name as an `OsString` instead,
keeping the bytes around as is.