fix: remove use of `from_utf8_unchecked` in macos process handling
Those unsafe functions have been observed to not respect their safety
constraints, as some values retrieved from the macos API are not
guaranteed to be UTF-8.

This is fixed by replacing those calls with `from_utf8_lossy`, which
will replace those values with a replacement character to ensure the
generated String is valid.

This solution is not ideal, but the master branch contains the better
fix: an OsString is returned to let the caller decide how to handle the
value.

Changing those calls remove the need of unsafe blocks, so those were
removed at the same time.